Models/5_report_generator_AI/requirements.txt
# FastAPI backend
fastapi
uvicorn[standard]

# LangChain ecosystem
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma

# Gemini (Google Generative AI)
google-generativeai
langchain-google-genai

# Vector DB (Chroma)
chromadb

# Embeddings (HuggingFace)
sentence-transformers
transformers
torch

# Data handling
pydantic
numpy
pandas

# Utils
python-dotenv
tqdm

# Logging
loguru